Try to reorganize the dialog
